Renaming to IsInActiveState
