+++ Specifications/Orleans/StateMachines/for_StateMachine/when_checking_if_is_in_active_state/and_it_is_not.cs
@@ -1,13 +1,13 @@
 // Copyright (c) Cratis. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Cratis.Applications.Orleans.StateMachines.when_checking_if_is_in_state;
+namespace Cratis.Applications.Orleans.StateMachines.when_checking_if_is_in_active_state;
 
 public class and_it_is_not : given.a_state_machine_with_well_known_states
 {
     bool result;
 
-    void Because() => result = state_machine.IsInState;
+    void Because() => result = state_machine.IsInActiveState;
 
     [Fact] void should_not_be_in_state() => result.ShouldBeFalse();
 }
